test: add unit tests for Poisson race likelihood
